@@ -10,7 +10,8 @@ Usage:
     credentials = CredentialManager()
     register_all_tools(mcp, credentials=credentials)
 """
-from typing import List, Optional, TYPE_CHECKING
+
+from typing import TYPE_CHECKING, Optional
 
 from fastmcp import FastMCP
 
@@ -18,26 +19,31 @@ if TYPE_CHECKING:
     from aden_tools.credentials import CredentialManager
 
 # Import register_tools from each tool module
+from .csv_tool import register_tools as register_csv
 from .example_tool import register_tools as register_example
-from .web_search_tool import register_tools as register_web_search
-from .web_scrape_tool import register_tools as register_web_scrape
-from .pdf_read_tool import register_tools as register_pdf_read
+from .file_system_toolkits.apply_diff import register_tools as register_apply_diff
+from .file_system_toolkits.apply_patch import register_tools as register_apply_patch
+from .file_system_toolkits.execute_command_tool import (
+    register_tools as register_execute_command,
+)
+from .file_system_toolkits.grep_search import register_tools as register_grep_search
+from .file_system_toolkits.list_dir import register_tools as register_list_dir
+from .file_system_toolkits.replace_file_content import (
+    register_tools as register_replace_file_content,
+)
 
 # Import file system toolkits
 from .file_system_toolkits.view_file import register_tools as register_view_file
 from .file_system_toolkits.write_to_file import register_tools as register_write_to_file
-from .file_system_toolkits.list_dir import register_tools as register_list_dir
-from .file_system_toolkits.replace_file_content import register_tools as register_replace_file_content
-from .file_system_toolkits.apply_diff import register_tools as register_apply_diff
-from .file_system_toolkits.apply_patch import register_tools as register_apply_patch
-from .file_system_toolkits.grep_search import register_tools as register_grep_search
-from .file_system_toolkits.execute_command_tool import register_tools as register_execute_command
+from .pdf_read_tool import register_tools as register_pdf_read
+from .web_scrape_tool import register_tools as register_web_scrape
+from .web_search_tool import register_tools as register_web_search
 
 
 def register_all_tools(
     mcp: FastMCP,
     credentials: Optional["CredentialManager"] = None,
-) -> List[str]:
+) -> list[str]:
     """
     Register all tools with a FastMCP server.
 
@@ -51,11 +57,11 @@ def register_all_tools(
     """
     # Tools that don't need credentials
     register_example(mcp)
-    register_web_search(mcp)
     register_web_scrape(mcp)
     register_pdf_read(mcp)
 
     # Tools that need credentials (pass credentials if provided)
+    # web_search supports multiple providers (Google, Brave) with auto-detection
     register_web_search(mcp, credentials=credentials)
 
     # Register file system toolkits
@@ -67,6 +73,7 @@ def register_all_tools(
     register_apply_patch(mcp)
     register_grep_search(mcp)
     register_execute_command(mcp)
+    register_csv(mcp)
 
     return [
         "example_tool",
@@ -81,6 +88,11 @@ def register_all_tools(
         "apply_patch",
         "grep_search",
         "execute_command_tool",
+        "csv_read",
+        "csv_write",
+        "csv_append",
+        "csv_info",
+        "csv_sql",
     ]
 
 
