@@ -13,11 +13,11 @@ Security measures:
 """
 
 import ast
-import sys
 import signal
-from typing import Any
-from dataclasses import dataclass, field
+import sys
 from contextlib import contextmanager
+from dataclasses import dataclass, field
+from typing import Any
 
 # Safe builtins whitelist
 SAFE_BUILTINS = {
@@ -25,7 +25,6 @@ SAFE_BUILTINS = {
     "True": True,
     "False": False,
     "None": None,
-
     # Type constructors
     "bool": bool,
     "int": int,
@@ -36,7 +35,6 @@ SAFE_BUILTINS = {
     "set": set,
     "tuple": tuple,
     "frozenset": frozenset,
-
     # Basic functions
     "abs": abs,
     "all": all,
@@ -97,22 +95,26 @@ BLOCKED_AST_NODES = {
 
 class CodeSandboxError(Exception):
     """Error during sandboxed code execution."""
+
     pass
 
 
 class TimeoutError(CodeSandboxError):
     """Code execution timed out."""
+
     pass
 
 
 class SecurityError(CodeSandboxError):
     """Code contains potentially dangerous operations."""
+
     pass
 
 
 @dataclass
 class SandboxResult:
     """Result of sandboxed code execution."""
+
     success: bool
     result: Any = None
     error: str | None = None
@@ -134,6 +136,7 @@ class RestrictedImporter:
 
         if name not in self._cache:
             import importlib
+
             self._cache[name] = importlib.import_module(name)
 
         return self._cache[name]
@@ -161,9 +164,8 @@ class CodeValidator:
         for node in ast.walk(tree):
             # Check for blocked node types
             if type(node) in self.blocked_nodes:
-                issues.append(
-                    f"Blocked operation: {type(node).__name__} at line {getattr(node, 'lineno', '?')}"
-                )
+                lineno = getattr(node, "lineno", "?")
+                issues.append(f"Blocked operation: {type(node).__name__} at line {lineno}")
 
             # Check for dangerous attribute access
             if isinstance(node, ast.Attribute):
@@ -212,11 +214,12 @@ class CodeSandbox:
     @contextmanager
     def _timeout_context(self, seconds: int):
         """Context manager for timeout enforcement."""
+
         def handler(signum, frame):
             raise TimeoutError(f"Code execution timed out after {seconds} seconds")
 
         # Only works on Unix-like systems
-        if hasattr(signal, 'SIGALRM'):
+        if hasattr(signal, "SIGALRM"):
             old_handler = signal.signal(signal.SIGALRM, handler)
             signal.alarm(seconds)
             try:
@@ -275,6 +278,7 @@ class CodeSandbox:
 
         # Capture stdout
         import io
+
         old_stdout = sys.stdout
         sys.stdout = captured_stdout = io.StringIO()
 
@@ -296,11 +300,7 @@ class CodeSandbox:
 
             # Also extract any new variables (not in inputs or builtins)
             for key, value in namespace.items():
-                if (
-                    key not in inputs
-                    and key not in self.safe_builtins
-                    and not key.startswith("_")
-                ):
+                if key not in inputs and key not in self.safe_builtins and not key.startswith("_"):
                     extracted[key] = value
 
             return SandboxResult(
