@@ -10,20 +10,24 @@ appropriate executor based on action type:
 - Code execution (sandboxed)
 """
 
-from typing import Any, Callable
-from dataclasses import dataclass, field
-import time
 import json
+import logging
 import re
+import time
+from collections.abc import Callable
+from dataclasses import dataclass, field
+from typing import Any
 
+from framework.graph.code_sandbox import CodeSandbox
 from framework.graph.plan import (
-    PlanStep,
     ActionSpec,
     ActionType,
+    PlanStep,
 )
-from framework.graph.code_sandbox import CodeSandbox
-from framework.runtime.core import Runtime
 from framework.llm.provider import LLMProvider, Tool
+from framework.runtime.core import Runtime
+
+logger = logging.getLogger(__name__)
 
 
 def parse_llm_json_response(text: str) -> tuple[Any | None, str]:
@@ -50,7 +54,7 @@ def parse_llm_json_response(text: str) -> tuple[Any | None, str]:
 
     # Try to extract JSON from markdown code blocks
     # Pattern: ```json ... ``` or ``` ... ```
-    code_block_pattern = r'```(?:json)?\s*([\s\S]*?)\s*```'
+    code_block_pattern = r"```(?:json)?\s*([\s\S]*?)\s*```"
     matches = re.findall(code_block_pattern, cleaned)
 
     if matches:
@@ -59,34 +63,46 @@ def parse_llm_json_response(text: str) -> tuple[Any | None, str]:
             try:
                 parsed = json.loads(match.strip())
                 return parsed, match.strip()
-            except json.JSONDecodeError:
+            except json.JSONDecodeError as e:
+                logger.debug(
+                    f"Failed to parse JSON from code block: {e}. "
+                    f"Content preview: {match.strip()[:100]}..."
+                )
                 continue
 
     # No code blocks or parsing failed - try parsing the whole response
     try:
         parsed = json.loads(cleaned)
         return parsed, cleaned
-    except json.JSONDecodeError:
-        pass
+    except json.JSONDecodeError as e:
+        logger.debug(
+            f"Failed to parse entire response as JSON: {e}. Content preview: {cleaned[:100]}..."
+        )
 
     # Try to find JSON-like content (starts with { or [)
-    json_start_pattern = r'(\{[\s\S]*\}|\[[\s\S]*\])'
+    json_start_pattern = r"(\{[\s\S]*\}|\[[\s\S]*\])"
     json_matches = re.findall(json_start_pattern, cleaned)
 
     for match in json_matches:
         try:
             parsed = json.loads(match)
             return parsed, match
-        except json.JSONDecodeError:
+        except json.JSONDecodeError as e:
+            logger.debug(f"Failed to parse JSON pattern: {e}. Content preview: {match[:100]}...")
             continue
 
-    # Could not parse as JSON
+    # Could not parse as JSON - log warning
+    logger.warning(
+        f"Could not parse LLM response as JSON after trying all strategies. "
+        f"Response preview: {cleaned[:200]}..."
+    )
     return None, cleaned
 
 
 @dataclass
 class StepExecutionResult:
     """Result of executing a plan step."""
+
     success: bool
     outputs: dict[str, Any] = field(default_factory=dict)
     error: str | None = None
@@ -160,11 +176,13 @@ class WorkerNode:
         # Record decision
         decision_id = self.runtime.decide(
             intent=f"Execute plan step: {step.description}",
-            options=[{
-                "id": step.action.action_type.value,
-                "description": f"Execute {step.action.action_type.value} action",
-                "action_type": step.action.action_type.value,
-            }],
+            options=[
+                {
+                    "id": step.action.action_type.value,
+                    "description": f"Execute {step.action.action_type.value} action",
+                    "action_type": step.action.action_type.value,
+                }
+            ],
             chosen=step.action.action_type.value,
             reasoning=f"Step requires {step.action.action_type.value}",
             context={"step_id": step.id, "inputs": step.inputs},
@@ -288,7 +306,7 @@ class WorkerNode:
             if inputs:
                 context_section = "\n\n--- Context Data ---\n"
                 for key, value in inputs.items():
-                    if isinstance(value, (dict, list)):
+                    if isinstance(value, dict | list):
                         context_section += f"{key}: {json.dumps(value, indent=2)}\n"
                     else:
                         context_section += f"{key}: {value}\n"
@@ -414,6 +432,7 @@ class WorkerNode:
         try:
             # Execute tool via formal executor
             from framework.llm.provider import ToolUse
+
             tool_use = ToolUse(
                 id=f"step_{tool_name}",
                 name=tool_name,
