@@ -1,7 +1,26 @@
 """LLM provider abstraction."""
 
 from framework.llm.provider import LLMProvider, LLMResponse
-from framework.llm.anthropic import AnthropicProvider
-from framework.llm.litellm import LiteLLMProvider
 
-__all__ = ["LLMProvider", "LLMResponse", "AnthropicProvider", "LiteLLMProvider"]
+__all__ = ["LLMProvider", "LLMResponse"]
+
+try:
+    from framework.llm.anthropic import AnthropicProvider  # noqa: F401
+
+    __all__.append("AnthropicProvider")
+except ImportError:
+    pass
+
+try:
+    from framework.llm.litellm import LiteLLMProvider  # noqa: F401
+
+    __all__.append("LiteLLMProvider")
+except ImportError:
+    pass
+
+try:
+    from framework.llm.mock import MockLLMProvider  # noqa: F401
+
+    __all__.append("MockLLMProvider")
+except ImportError:
+    pass
