@@ -1,6 +1,7 @@
 """LLM Provider abstraction for pluggable LLM backends."""
 
 from abc import ABC, abstractmethod
+from collections.abc import Callable
 from dataclasses import dataclass, field
 from typing import Any
 
@@ -8,6 +9,7 @@ from typing import Any
 @dataclass
 class LLMResponse:
     """Response from an LLM call."""
+
     content: str
     model: str
     input_tokens: int = 0
@@ -19,6 +21,7 @@ class LLMResponse:
 @dataclass
 class Tool:
     """A tool the LLM can use."""
+
     name: str
     description: str
     parameters: dict[str, Any] = field(default_factory=dict)
@@ -27,6 +30,7 @@ class Tool:
 @dataclass
 class ToolUse:
     """A tool call requested by the LLM."""
+
     id: str
     name: str
     input: dict[str, Any]
@@ -35,6 +39,7 @@ class ToolUse:
 @dataclass
 class ToolResult:
     """Result of executing a tool."""
+
     tool_use_id: str
     content: str
     is_error: bool = False
@@ -58,6 +63,8 @@ class LLMProvider(ABC):
         system: str = "",
         tools: list[Tool] | None = None,
         max_tokens: int = 1024,
+        response_format: dict[str, Any] | None = None,
+        json_mode: bool = False,
     ) -> LLMResponse:
         """
         Generate a completion from the LLM.
@@ -67,6 +74,11 @@ class LLMProvider(ABC):
             system: System prompt
             tools: Available tools for the LLM to use
             max_tokens: Maximum tokens to generate
+            response_format: Optional structured output format. Use:
+                - {"type": "json_object"} for basic JSON mode
+                - {"type": "json_schema", "json_schema": {"name": "...", "schema": {...}}}
+                  for strict JSON schema enforcement
+            json_mode: If True, request structured JSON output from the LLM
 
         Returns:
             LLMResponse with content and metadata
@@ -79,7 +91,7 @@ class LLMProvider(ABC):
         messages: list[dict[str, Any]],
         system: str,
         tools: list[Tool],
-        tool_executor: callable,
+        tool_executor: Callable[["ToolUse"], "ToolResult"],
         max_iterations: int = 10,
     ) -> LLMResponse:
         """
