@@ -6,6 +6,7 @@ Supports both STDIO and HTTP transports using the official MCP Python SDK.
 
 import asyncio
 import logging
+import os
 from dataclasses import dataclass, field
 from typing import Any, Literal
 
@@ -145,13 +146,16 @@ class MCPClient:
 
         try:
             import threading
+
             from mcp import StdioServerParameters
 
             # Create server parameters
+            # Always inherit parent environment and merge with any custom env vars
+            merged_env = {**os.environ, **(self.config.env or {})}
             server_params = StdioServerParameters(
                 command=self.config.command,
                 args=self.config.args,
-                env=self.config.env or None,
+                env=merged_env,
                 cwd=self.config.cwd,
             )
 
@@ -177,7 +181,10 @@ class MCPClient:
 
                         # Create persistent stdio client context
                         self._stdio_context = stdio_client(server_params)
-                        self._read_stream, self._write_stream = await self._stdio_context.__aenter__()
+                        (
+                            self._read_stream,
+                            self._write_stream,
+                        ) = await self._stdio_context.__aenter__()
 
                         # Create persistent session
                         self._session = ClientSession(self._read_stream, self._write_stream)
@@ -212,7 +219,7 @@ class MCPClient:
 
             logger.info(f"Connected to MCP server '{self.config.name}' via STDIO (persistent)")
         except Exception as e:
-            raise RuntimeError(f"Failed to connect to MCP server: {e}")
+            raise RuntimeError(f"Failed to connect to MCP server: {e}") from e
 
     def _connect_http(self) -> None:
         """Connect to MCP server via HTTP transport."""
@@ -229,7 +236,9 @@ class MCPClient:
         try:
             response = self._http_client.get("/health")
             response.raise_for_status()
-            logger.info(f"Connected to MCP server '{self.config.name}' via HTTP at {self.config.url}")
+            logger.info(
+                f"Connected to MCP server '{self.config.name}' via HTTP at {self.config.url}"
+            )
         except Exception as e:
             logger.warning(f"Health check failed for MCP server '{self.config.name}': {e}")
             # Continue anyway, server might not have health endpoint
@@ -252,7 +261,10 @@ class MCPClient:
                 )
                 self._tools[tool.name] = tool
 
-            logger.info(f"Discovered {len(self._tools)} tools from '{self.config.name}': {list(self._tools.keys())}")
+            tool_names = list(self._tools.keys())
+            logger.info(
+                f"Discovered {len(self._tools)} tools from '{self.config.name}': {tool_names}"
+            )
         except Exception as e:
             logger.error(f"Failed to discover tools from '{self.config.name}': {e}")
             raise
@@ -268,11 +280,13 @@ class MCPClient:
         # Convert tools to dict format
         tools_list = []
         for tool in response.tools:
-            tools_list.append({
-                "name": tool.name,
-                "description": tool.description,
-                "inputSchema": tool.inputSchema,
-            })
+            tools_list.append(
+                {
+                    "name": tool.name,
+                    "description": tool.description,
+                    "inputSchema": tool.inputSchema,
+                }
+            )
 
         return tools_list
 
@@ -300,7 +314,7 @@ class MCPClient:
 
             return data.get("result", {}).get("tools", [])
         except Exception as e:
-            raise RuntimeError(f"Failed to list tools via HTTP: {e}")
+            raise RuntimeError(f"Failed to list tools via HTTP: {e}") from e
 
     def list_tools(self) -> list[MCPTool]:
         """
@@ -350,9 +364,9 @@ class MCPClient:
             if len(result.content) > 0:
                 content_item = result.content[0]
                 # Check if it's a text content item
-                if hasattr(content_item, 'text'):
+                if hasattr(content_item, "text"):
                     return content_item.text
-                elif hasattr(content_item, 'data'):
+                elif hasattr(content_item, "data"):
                     return content_item.data
             return result.content
 
@@ -384,7 +398,7 @@ class MCPClient:
 
             return data.get("result", {}).get("content", [])
         except Exception as e:
-            raise RuntimeError(f"Failed to call tool via HTTP: {e}")
+            raise RuntimeError(f"Failed to call tool via HTTP: {e}") from e
 
     def disconnect(self) -> None:
         """Disconnect from the MCP server."""
