@@ -1,10 +1,10 @@
 """Message protocol for multi-agent communication."""
 
+import uuid
 from dataclasses import dataclass, field
 from datetime import datetime
 from enum import Enum
 from typing import Any
-import uuid
 
 
 class MessageType(Enum):
