@@ -6,10 +6,10 @@ storing tests as JSON files with indexes for efficient querying.
 """
 
 import json
-from pathlib import Path
 from datetime import datetime
+from pathlib import Path
 
-from framework.testing.test_case import Test, ApprovalStatus, TestType
+from framework.testing.test_case import ApprovalStatus, Test, TestType
 from framework.testing.test_result import TestResult
 
 
@@ -34,6 +34,8 @@ class TestStorage:
       suites/
         {goal_id}_suite.json       # Test suite metadata
     """
+
+    __test__ = False  # Not a pytest test class
 
     def __init__(self, base_path: str | Path):
         self.base_path = Path(base_path)
@@ -197,8 +199,7 @@ class TestStorage:
 
         # Get all result files except latest.json
         result_files = sorted(
-            [f for f in results_dir.glob("*.json") if f.name != "latest.json"],
-            reverse=True
+            [f for f in results_dir.glob("*.json") if f.name != "latest.json"], reverse=True
         )[:limit]
 
         results = []
