@@ -6,44 +6,59 @@ This script checks if the MCP server is properly installed and configured.
 """
 
 import json
+import logging
 import subprocess
 import sys
 from pathlib import Path
 
+logger = logging.getLogger(__name__)
+
+
+def setup_logger():
+    """Configure logger for CLI usage."""
+    if not logger.handlers:
+        handler = logging.StreamHandler(sys.stdout)
+        formatter = logging.Formatter("%(message)s")
+        handler.setFormatter(formatter)
+        logger.addHandler(handler)
+        logger.setLevel(logging.INFO)
+
 
 class Colors:
-    GREEN = '\033[0;32m'
-    YELLOW = '\033[1;33m'
-    RED = '\033[0;31m'
-    BLUE = '\033[0;34m'
-    NC = '\033[0m'
+    GREEN = "\033[0;32m"
+    YELLOW = "\033[1;33m"
+    RED = "\033[0;31m"
+    BLUE = "\033[0;34m"
+    NC = "\033[0m"
 
 
 def check(description: str) -> bool:
     """Print check description and return a context manager for result."""
-    print(f"Checking {description}...", end=" ")
+    logger.info(f"Checking {description}... ", extra={"end": ""})
+    sys.stdout.flush()
     return True
 
 
 def success(msg: str = "OK"):
-    """Print success message."""
-    print(f"{Colors.GREEN}✓ {msg}{Colors.NC}")
+    """Log success message."""
+    logger.info(f"{Colors.GREEN}✓ {msg}{Colors.NC}")
 
 
 def warning(msg: str):
-    """Print warning message."""
-    print(f"{Colors.YELLOW}⚠ {msg}{Colors.NC}")
+    """Log warning message."""
+    logger.warning(f"{Colors.YELLOW}⚠ {msg}{Colors.NC}")
 
 
 def error(msg: str):
-    """Print error message."""
-    print(f"{Colors.RED}✗ {msg}{Colors.NC}")
+    """Log error message."""
+    logger.error(f"{Colors.RED}✗ {msg}{Colors.NC}")
 
 
 def main():
     """Run verification checks."""
-    print("=== MCP Server Verification ===")
-    print()
+    setup_logger()
+    logger.info("=== MCP Server Verification ===")
+    logger.info("")
 
     script_dir = Path(__file__).parent.absolute()
     all_checks_passed = True
@@ -55,13 +70,13 @@ def main():
             [sys.executable, "-c", "import framework; print(framework.__file__)"],
             capture_output=True,
             text=True,
-            check=True
+            check=True,
         )
         framework_path = result.stdout.strip()
         success(f"installed at {framework_path}")
     except subprocess.CalledProcessError:
         error("framework package not found")
-        print(f"  Run: pip install -e {script_dir}")
+        logger.info(f"  Run: pip install -e {script_dir}")
         all_checks_passed = False
 
     # Check 2: MCP dependencies
@@ -69,17 +84,13 @@ def main():
     missing_deps = []
     for dep in ["mcp", "fastmcp"]:
         try:
-            subprocess.run(
-                [sys.executable, "-c", f"import {dep}"],
-                capture_output=True,
-                check=True
-            )
+            subprocess.run([sys.executable, "-c", f"import {dep}"], capture_output=True, check=True)
         except subprocess.CalledProcessError:
             missing_deps.append(dep)
 
     if missing_deps:
         error(f"missing: {', '.join(missing_deps)}")
-        print(f"  Run: pip install {' '.join(missing_deps)}")
+        logger.info(f"  Run: pip install {' '.join(missing_deps)}")
         all_checks_passed = False
     else:
         success("all installed")
@@ -91,12 +102,12 @@ def main():
             [sys.executable, "-c", "from framework.mcp import agent_builder_server"],
             capture_output=True,
             text=True,
-            check=True
+            check=True,
         )
         success("loads successfully")
     except subprocess.CalledProcessError as e:
         error("failed to import")
-        print(f"  Error: {e.stderr}")
+        logger.error(f"  Error: {e.stderr}")
         all_checks_passed = False
 
     # Check 4: MCP configuration file
@@ -110,9 +121,9 @@ def main():
             if "mcpServers" in config and "agent-builder" in config["mcpServers"]:
                 server_config = config["mcpServers"]["agent-builder"]
                 success("found and valid")
-                print(f"  Command: {server_config.get('command')}")
-                print(f"  Args: {' '.join(server_config.get('args', []))}")
-                print(f"  CWD: {server_config.get('cwd')}")
+                logger.info(f"  Command: {server_config.get('command')}")
+                logger.info(f"  Args: {' '.join(server_config.get('args', []))}")
+                logger.info(f"  CWD: {server_config.get('cwd')}")
             else:
                 warning("exists but missing agent-builder config")
                 all_checks_passed = False
@@ -121,8 +132,8 @@ def main():
             all_checks_passed = False
     else:
         warning("not found (optional)")
-        print(f"  Location would be: {mcp_config}")
-        print("  Run setup_mcp.py to create it")
+        logger.info(f"  Location would be: {mcp_config}")
+        logger.info("  Run setup_mcp.py to create it")
 
     # Check 5: Framework modules
     check("core framework modules")
@@ -138,9 +149,7 @@ def main():
     for module in modules_to_check:
         try:
             subprocess.run(
-                [sys.executable, "-c", f"import {module}"],
-                capture_output=True,
-                check=True
+                [sys.executable, "-c", f"import {module}"], capture_output=True, check=True
             )
         except subprocess.CalledProcessError:
             failed_modules.append(module)
@@ -156,12 +165,15 @@ def main():
     try:
         # Try to import and instantiate the MCP server
         result = subprocess.run(
-            [sys.executable, "-c",
-             "from framework.mcp.agent_builder_server import mcp; print('OK')"],
+            [
+                sys.executable,
+                "-c",
+                "from framework.mcp.agent_builder_server import mcp; print('OK')",
+            ],
             capture_output=True,
             text=True,
             check=True,
-            timeout=5
+            timeout=5,
         )
         if "OK" in result.stdout:
             success("server can start")
@@ -171,28 +183,28 @@ def main():
         warning("server startup slow (might be OK)")
     except subprocess.CalledProcessError as e:
         error("server failed to start")
-        print(f"  Error: {e.stderr}")
+        logger.error(f"  Error: {e.stderr}")
         all_checks_passed = False
 
-    print()
-    print("=" * 40)
+    logger.info("")
+    logger.info("=" * 40)
     if all_checks_passed:
-        print(f"{Colors.GREEN}✓ All checks passed!{Colors.NC}")
-        print()
-        print("Your MCP server is ready to use.")
-        print()
-        print(f"{Colors.BLUE}To start the server:{Colors.NC}")
-        print("  python -m framework.mcp.agent_builder_server")
-        print()
-        print(f"{Colors.BLUE}To use with Claude Desktop:{Colors.NC}")
-        print("  Add the configuration from .mcp.json to your")
-        print("  Claude Desktop MCP settings")
+        logger.info(f"{Colors.GREEN}✓ All checks passed!{Colors.NC}")
+        logger.info("")
+        logger.info("Your MCP server is ready to use.")
+        logger.info("")
+        logger.info(f"{Colors.BLUE}To start the server:{Colors.NC}")
+        logger.info("  python -m framework.mcp.agent_builder_server")
+        logger.info("")
+        logger.info(f"{Colors.BLUE}To use with Claude Desktop:{Colors.NC}")
+        logger.info("  Add the configuration from .mcp.json to your")
+        logger.info("  Claude Desktop MCP settings")
     else:
-        print(f"{Colors.RED}✗ Some checks failed{Colors.NC}")
-        print()
-        print("To fix issues, run:")
-        print(f"  python {script_dir / 'setup_mcp.py'}")
-    print()
+        logger.info(f"{Colors.RED}✗ Some checks failed{Colors.NC}")
+        logger.info("")
+        logger.info("To fix issues, run:")
+        logger.info(f"  python {script_dir / 'setup_mcp.py'}")
+    logger.info("")
 
 
 if __name__ == "__main__":
