@@ -25,39 +25,59 @@ Note:
     - Tier 2 (agent load): Tool credentials validated when agent is loaded
     See aden_tools.credentials for details.
 """
+
 import argparse
+import logging
 import os
 import sys
+
+logger = logging.getLogger(__name__)
+
+
+def setup_logger():
+    """Configure logger for MCP server."""
+    if not logger.handlers:
+        # For STDIO mode, log to stderr; for HTTP mode, log to stdout
+        stream = sys.stderr if "--stdio" in sys.argv else sys.stdout
+        handler = logging.StreamHandler(stream)
+        formatter = logging.Formatter("[MCP] %(message)s")
+        handler.setFormatter(formatter)
+        logger.addHandler(handler)
+        logger.setLevel(logging.INFO)
+
+
+setup_logger()
 
 # Suppress FastMCP banner in STDIO mode
 if "--stdio" in sys.argv:
     # Monkey-patch rich Console to redirect to stderr
     import rich.console
+
     _original_console_init = rich.console.Console.__init__
 
     def _patched_console_init(self, *args, **kwargs):
-        kwargs['file'] = sys.stderr  # Force all rich output to stderr
+        kwargs["file"] = sys.stderr  # Force all rich output to stderr
         _original_console_init(self, *args, **kwargs)
 
     rich.console.Console.__init__ = _patched_console_init
 
-from fastmcp import FastMCP
-from starlette.requests import Request
-from starlette.responses import PlainTextResponse
+from fastmcp import FastMCP  # noqa: E402
+from starlette.requests import Request  # noqa: E402
+from starlette.responses import PlainTextResponse  # noqa: E402
 
-from aden_tools.credentials import CredentialManager, CredentialError
-from aden_tools.tools import register_all_tools
+from aden_tools.credentials import CredentialError, CredentialManager  # noqa: E402
+from aden_tools.tools import register_all_tools  # noqa: E402
 
 # Create credential manager
 credentials = CredentialManager()
 
-# Tier 1: Validate startup-required credentials (ANTHROPIC_API_KEY)
+# Tier 1: Validate startup-required credentials (if any)
 try:
     credentials.validate_startup()
-    print("[MCP] Startup credentials validated")
+    logger.info("Startup credentials validated")
 except CredentialError as e:
-    print(f"[MCP] FATAL: {e}", file=sys.stderr)
-    sys.exit(1)
+    # Non-fatal - tools will validate their own credentials when called
+    logger.warning(str(e))
 
 mcp = FastMCP("tools")
 
@@ -65,7 +85,7 @@ mcp = FastMCP("tools")
 tools = register_all_tools(mcp, credentials=credentials)
 # Only print to stdout in HTTP mode (STDIO mode requires clean stdout for JSON-RPC)
 if "--stdio" not in sys.argv:
-    print(f"[MCP] Registered {len(tools)} tools: {tools}")
+    logger.info(f"Registered {len(tools)} tools: {tools}")
 
 
 @mcp.custom_route("/health", methods=["GET"])
@@ -105,7 +125,7 @@ def main() -> None:
         # STDIO mode: only JSON-RPC messages go to stdout
         mcp.run(transport="stdio")
     else:
-        print(f"[MCP] Starting HTTP server on {args.host}:{args.port}")
+        logger.info(f"Starting HTTP server on {args.host}:{args.port}")
         mcp.run(transport="http", host=args.host, port=args.port)
 
 
