@@ -3,6 +3,7 @@ LLM provider credentials.
 
 Contains credentials for language model providers like Anthropic, OpenAI, etc.
 """
+
 from .base import CredentialSpec
 
 LLM_CREDENTIALS = {
@@ -10,10 +11,10 @@ LLM_CREDENTIALS = {
         env_var="ANTHROPIC_API_KEY",
         tools=[],
         node_types=["llm_generate", "llm_tool_use"],
-        required=True,
-        startup_required=True,
+        required=False,  # Not required - agents can use other providers via LiteLLM
+        startup_required=False,  # MCP server doesn't need LLM credentials
         help_url="https://console.anthropic.com/settings/keys",
-        description="API key for Anthropic Claude models (required for testing)",
+        description="API key for Anthropic Claude models",
     ),
     # Future LLM providers:
     # "openai": CredentialSpec(
