@@ -1,7 +1,8 @@
 """Shared fixtures for tools tests."""
-import pytest
+
 from pathlib import Path
 
+import pytest
 from fastmcp import FastMCP
 
 from aden_tools.credentials import CredentialManager
@@ -16,11 +17,13 @@ def mcp() -> FastMCP:
 @pytest.fixture
 def mock_credentials() -> CredentialManager:
     """Create a CredentialManager with mock test credentials."""
-    return CredentialManager.for_testing({
-        "anthropic": "test-anthropic-api-key",
-        "brave_search": "test-brave-api-key",
-        # Add other mock credentials as needed
-    })
+    return CredentialManager.for_testing(
+        {
+            "anthropic": "test-anthropic-api-key",
+            "brave_search": "test-brave-api-key",
+            # Add other mock credentials as needed
+        }
+    )
 
 
 @pytest.fixture
