@@ -1,7 +1,8 @@
 """Tests for web_scrape tool (FastMCP)."""
-import pytest
 
+import pytest
 from fastmcp import FastMCP
+
 from aden_tools.tools.web_scrape_tool import register_tools
 
 
@@ -50,3 +51,12 @@ class TestWebScrapeTool:
         """selector parameter is accepted."""
         result = web_scrape_fn(url="https://example.com", selector=".content")
         assert isinstance(result, dict)
+
+    def test_non_html_content_rejected(self, web_scrape_fn):
+        """Ensure non-HTML content types (like JSON) are rejected."""
+        # GitHub's Zen API returns text/plain, not html
+        result = web_scrape_fn(url="https://api.github.com/zen")
+
+        # We expect an error about skipping non-HTML
+        assert "error" in result
+        assert "Skipping non-HTML content" in result["error"]
